Use floats for caret drawing.

+++ src/gra/proto.h
@@ -299,7 +299,7 @@ COMMON(status)	makeClassScrollBar(Class class);
 /* /swi40/jan/src/pl/packages/xpce/src/gra/text.c */
 COMMON(void)	str_format(PceString out, const PceString in, const int width, const FontObj font);
 COMMON(void)	str_one_line(PceString to, PceString from);
-COMMON(void)	draw_caret(int x, int y, int w, int h, bool active);
+COMMON(void)	draw_caret(double x, double y, double w, double h, bool active);
 COMMON(status)	repaintText(TextObj t, int x, int y, int w, int h);
 COMMON(Int)	get_pointed_text(TextObj t, int x, int y);
 COMMON(status)	transparentText(TextObj t, BoolObj val);
